+++ activerecord/lib/active_record/secure_token.rb
@@ -50,7 +50,10 @@ module ActiveRecord
       #   An optional string prepended to the generated token. Intended primarily for keys that
       #   will be publicly visible (e.g. API keys, URLs, etc.) so that their purpose can be identified at a glance.
       #   The prefix does not count toward +:length+.
-      def has_secure_token(attribute = :token, length: MINIMUM_TOKEN_LENGTH, on: ActiveRecord.generate_secure_token_on, prefix: nil)
+      #
+      # [+:skip_inspection_filter]
+      #   Skip adding the column to the models <tt>filter_attributes</tt>. Defaults to false.
+      def has_secure_token(attribute = :token, length: MINIMUM_TOKEN_LENGTH, on: ActiveRecord.generate_secure_token_on, prefix: nil, skip_inspection_filter: false)
         if length < MINIMUM_TOKEN_LENGTH
           raise MinimumLengthError, "Token requires a minimum length of #{MINIMUM_TOKEN_LENGTH} characters."
         end
@@ -67,6 +70,10 @@ module ActiveRecord
           generate_token = -> { self.generate_unique_secure_token(length: length) }
         end
 
+        unless skip_inspection_filter
+          self.filter_attributes += [secure_token_filter_attribute(attribute, prefix)]
+        end
+
         # Load securerandom only when has_secure_token is used.
         require "active_support/core_ext/securerandom"
         define_method("regenerate_#{attribute}") { update! attribute => generate_token.call }
@@ -80,6 +87,19 @@ module ActiveRecord
       def generate_unique_secure_token(length: MINIMUM_TOKEN_LENGTH)
         SecureRandom.base58(length)
       end
+
+      private
+        def secure_token_filter_attribute(attribute, prefix)
+          return attribute unless prefix
+
+          ->(key, value) do
+            return unless key == attribute.to_s
+
+            value.replace(
+              "#{prefix}#{ActiveSupport::ParameterFilter::FILTERED}"
+            )
+          end
+        end
     end
   end
 end